// src/models/servicios.ts
export class Servicios {

  id: string;
  idExpediente: string;
  tipo: string; //A (AEREO) - M (MARITIMO)
  despachoAereo: boolean;
  despachoMaritimoExport: boolean;
  despachoMaritimoImport: boolean;
  revisionParaduanera: boolean;
  revisionFisica: boolean;
  t2l: boolean;
  ata: boolean;
  dae: boolean;
  eur1: boolean;
  entregaDocumentacion: boolean;
  recogidaDocumentos: boolean;

}